+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.2.8 (unreleased)
 
 - Correctly populate system-level variables when attaching existing (factory default) feature templates
+- Add `sdwan_intrusion_prevention_policy_definition` resource and data source
 
 ## 0.2.7